+++ include/kernel/core/utils.h
@@ -8,6 +8,6 @@ typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
 typedef unsigned long long uint64_t;
-typedef unsigned long long size_t;
+typedef unsigned long long uint64_t;
 
-char* convert(size_t num, size_t base);
+char* convert(uint64_t num, uint64_t base);
